#304 refactor(routine): Change alarm ID type from Long to Int
This commit refactors the `idAlarm` property in routine-related models to use `Int` instead of `Long`, standardizing the identifier type across the application layers.

This change is propagated through the domain model, presentation models, and the `WakeupActivity`, ensuring consistent type handling for alarm IDs.

**Key Changes:**

*   **`RoutineUiModel.kt`** (home & routine presentation):
    *   Changed the `idAlarm` property type from `Long?` to `Int?`.
*   **`Routine.kt`**:
    *   Changed the `idAlarm` property type from `Long?` to `Int?`.
*   **`WakeupActivity.kt`**:
    *   Updated the `routineId` property type to `Int?`.
    *   Replaced `getLongExtra` with `getIntExtra` when retrieving the alarm ID from the intent.

+++ app/src/main/kotlin/com/rahim/ui/wakeup/WakeupActivity.kt
@@ -49,7 +49,7 @@ class WakeupActivity : ComponentActivity() {
 
   private val controlAlarm: ControlAlarm by inject()
   private var routineName: String? = null
-  private var routineId: Long? = null
+  private var routineId: Int? = null
 
   override fun onCreate(savedInstanceState: Bundle?) {
     installSplashScreen()
@@ -131,7 +131,7 @@ class WakeupActivity : ComponentActivity() {
 
   private fun getIntentResult() {
     routineName = intent.getStringExtra(Constants.KEY_LAUNCH_NAME)
-    routineId = intent.getLongExtra(Constants.KEY_REMINDER_ALARM_ID, 0)
+    routineId = intent.getIntExtra(Constants.KEY_REMINDER_ALARM_ID, 0)
   }
 
   private fun setWakeupSetting() {
